feat: Modularize authentication system and enhance UI components
✨ New Features:
- Created modular authentication architecture by breaking down monolithic Signup.jsx
- Added stylish RoleDropdown component with modern animations and icons
- Implemented centered authentication layout with improved UX

🏗️ Component Restructuring:
- Split Authentication.jsx into separate Login, Signup, and ForgotPassword components
- Created reusable PasswordToggleIcon and SocialLogin components
- Extracted RoleDropDown as standalone component with enhanced styling

🎨 UI/UX Improvements:
- Added modern dropdown with smooth animations and hover effects
- Implemented color-coded role icons (Farmer, Officer, Admin, Subsidy Provider)
- Enhanced authentication form centering and responsive design
- Added stylish Back button with hover animations
- Fixed Login button to stay green during forgot password flow

📦 Dependencies:
- Added react-icons for modern SVG icons
- Added FontAwesome packages for additional icon options
- Updated package.json and package-lock.json

🐛 Bug Fixes:
- Fixed React component rendering issues in RoleDropdown
- Resolved JSX syntax errors in Authentication component
- Fixed navigation issues with homepage button

🔄 Code Organization:
- Maintained all original functionality while improving maintainability
- Enhanced code readability and component separation
- Preserved existing API integrations and authentication flows

All design and functionality remain unchanged while significantly improving code structure and user experience.
